extract Milc::Command from Milc::Base
